+++ client/src/components/Header.module.css
@@ -59,7 +59,7 @@
     margin: 0 8px;
   }
 
-  button {
+  > button {
     width: 40px;
     height: 40px;
   }
@@ -110,7 +110,8 @@
         margin: 0;
       }
 
-      > a {
+      > a,
+      button {
         font-size: $font5;
         font-weight: $font5W;
       }
