+++ client/src/components/modal/AdminModal.module.css
@@ -1,4 +1,4 @@
-@import url('../default.css');
+@import url('../../default.css');
 
 .backdrop {
   z-index: 2;
